eyecare_backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & Deep Learning
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Database - Using MongoDB
# MongoDB
pymongo==4.6.0
motor==3.3.2
beanie==1.24.0

# Data Visualization & Analysis
matplotlib==3.8.2
seaborn==0.13.0

# System Monitoring
psutil==5.9.6

# Utilities
python-dateutil==2.8.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For GPU support (uncomment if you have NVIDIA GPU)
# tensorflow-gpu==2.15.0
